Projet avacement

+++ ValaisEat/DTO/Order.cs
@@ -10,7 +10,7 @@ namespace DTO
         public string Status { get; set; }
         public DateTime Date { get; set; }
         public DateTime ShippingDate { get; set; }
-        public float TotalPrice { get; set; }
+        public double TotalPrice { get; set; }
         public int IdCourier { get; set; }
         public int IdClient { get; set; }
